scm: move FileNotInRepoError and MergeConflictError to scm.exceptions
