+++ dvc/repo/experiments/apply.py
@@ -21,7 +21,8 @@ logger = logging.getLogger(__name__)
 @scm_context
 def apply(repo, rev, force=True, **kwargs):
     from dvc.repo.checkout import checkout as dvc_checkout
-    from dvc.scm.base import MergeConflictError, SCMError
+    from dvc.scm.base import SCMError
+    from dvc.scm.exceptions import MergeConflictError
 
     exps = repo.experiments
